+++ ChangeLog.md
@@ -24,6 +24,7 @@
   - Zeige Gleise direkt nach dem hinzufügen an.
   - Korrigiere Offset bei gedrehtem Canvas.
   - Anpassen des Mauszeigers, wenn er über einem Gleis ist.
+  - Anpassen des Mauszeigers, wenn er über einem `Knopf` ist und dadurch ein Gleis hinzugefügt wurde.
   - Bevorzuge genau angeklickte Gleise, bevor der Toleranz-Bereich berücksichtigt wird.
 - Ersetzte Debug-Ausgaben für Fehler durch Display-Ausgabe über `thiserror::Error`
 - Aktualisiere auf iced v0.12